Fix too aggressive cache clearing when triggering an autopublish event
Previously the autopublish would clear all default cache providers, including context_settings for all contexts, system_settings, lexicons, resources etc. With this patch, it brings the cache clear in line with what happens when you save a resource, by only clearing the relevant context_settings (because the aliasMap needs to be updated), resource, and the db cache.

The reason this change is high-traffic sites with lots of contexts. The particular client I worked for this on has 300 (o.O) contexts with probably over 100.000 resources. When a front-end request triggers an auto publish, it would literally take the site down for a couple minutes as it regenerates all context caches (looping over all resources in all contexts to generate the aliasMap). Because multiple concurrent requests would hit the site, they'd all get stuck trying to generate those caches causing the error log to fill up with caching errors because they can't get a file lock (as another request is already doing the same thing).

It's been tricky to test this in a simulated environment (I couldn't get enough traffic to hit my local dev server to consistently trigger the problem), but the fix has been verified on the client production site. Where previously an auto publish would cause the server load to go from 0.4 to 70 and it would take 5-10 minutes for the site to come back online, with the fix applied there has not been a noticeable increase when a page gets (un)published making it as quick as ever.