Add $modx->deprecated method to log usage of deprecated methods
This is to be used as a standardised way to mark functionality as deprecated, by calling it from within the deprecated function with at least the version it was marked as deprecated in. Additional information like the calling function is automatically added, to ease identifying code that needs to be updated.

If the deprecated function cannot be automatically determined properly (e.g. the deprecated function is a snippet, which would show up as "include" in "modScript" being deprecated, or if a deprecation pertains about the usage of a full object) the third param can be provided to manually name the deprecated function.