Improve deprecated notice for "Old modTemplateVar getRender input"
Aside from changing the wording and adding context to make it a bit clearer what's the problem, this also moves it into the conditional. Before, the message would get logged a lot for each first TV of a certain type due to the way the foreach handles registered render methods. Now it only gets logged when a deprecated flat file is used.