Handle deprecated $type and $responseCode parameters in $modx->sendRedirect and fix message
When calling $modx->sendRedirect with the $type / $responseCode parameter set (instead of the values in an options array), the deprecation error message would look like this:

    [2019-01-11 21:42:21] (ERROR in modX::sendRedirect @ core/model/modx/modx.class.php : 1139) sendRedirect method parameter $type is deprecated since version 2.0.5. Use responseCode in options array instead.

Most notably, the message says it was triggered from modX::sendRedirect, while modX::deprecated is supposed to show you where it was called from. That's because the deprecated handling is being done in modResponse->sendRedirect, so the "offending" call is the `$this->response->sendRedirect` call. While accurate, it's not helpful in tracking down the source of the sendRedirect call.

With these changes, the modX::sendRedirect method handles the same deprecation notices. So any offending caller to modX::sendRedirect will get logged with a useful message. By also moving the $type and $responseCode into the options array, it prevents modResponse::sendRedirect from triggering the deprecated log again.

Also minor error in the deprecation message in modResponse has been fixed.