Refactor template variable creation (#15773)
* Refactor of TV Configuration and Management

A continuation of the work initiated in draft PR #15164, ported to 3.x. Many of the validation UX features (and more) that were originally in the referenced draft (and have since been implemented into 2.x via PR #15146) have been removed from this commit and pending PR for clarity.

* GIT ignore

Ignore this file, had to commit to reapply stashed changes.

* Updates to Pass Linter

Also a couple minor tweaks were made.

* Requested changes from PR conversation, batch 1

* A couple more minor changes

* Update resourcelist.tpl

* Changes in Response to PR Feedback

Addresses comments and requests to-date, with the exception of documentation.

* PSR-12 Fixes

* Revert gitignore changes

* Changes in Response to PR Feedback, Batch 2

* Updates to Enable Multiple Overrides

A few additions to allow non-native tvs to switch to the new loader method and add their own named validators and listeners.

Note that I originally thought it would be possible for each custom TV to provide its own Ext.override(MODx.panel.TV,{}), and that overrides would build on top of one another. That's not the case however, as the last-loaded Ext.override takes precedence over all others.

* grunt build

Co-authored-by: Jason Coward <jason@opengeek.com>