Add missing system setting for profile photo media source (#16515)
### What does it do?
Adds a long-missing setting for specifying user's profile photo media
source. Also, moves up (and tweaks) the associated Lexicon entries.

### Why is it needed?
The logic has been present in the backend to specify a Media Source
other that the default for user profile photos since they were
introduced nearly 10 years ago. However, the system setting was never
added to the build, making it virtually unknown that this functionality
existed.

### How to test
Run `_build/transport.core.php` to install the new setting. Then search
for `photo_profile_source` in the system settings, change its value, and
verify that the Media Source shown in the Media Browser (when
editing/adding a user profile photo) matches that which you specified in
the setting.

### Related issue(s)/PR(s)
Resolves #16510