Fix Dashboard Updates widget's display of the most recent modx version (#16608)
### What does it do?
Added new `SoftwareUpdate` Processors for retrieving available MODX and
Extras updates information, utilizing the new upgrades API. This PR is a
Phase 1 fix that solves the issue at hand, but also paves the way for a
more flexible and comprehensive Updates widget display to be done in a
Phase 2 PR. In that next phase, users will be able to control what level
of update MODX candidates should be shown (via a few new system
settings).

The new MODX download functionality provides a direct download instead
of a link to the modx.com downloads area.

### Why is it needed?
The widget currently incorrectly shows the current -dev version.

### How to test

1. Manually delete session data from the database.
2. Clear browser and MODX caches and verify that the widget displays the
correct update version.
3. Test with an up-to-date `-pl` release (currently 3.0.5), a `-dev`
release (github 3.0.x or 3.x), and an older `-pl` or `-dev` release
(3.0.0 - 3.0.4). This is a little tricky, because you're going to pull
down the 3.1.0-dev from github to review. You may want to manually
override the installed version in
`core/src/Revolution/Processors/SoftwareUpdate/GetList.php` (line 33),
setting it to various versions to see what happens in the Dashboard
widget (e.g., 3.0.1-pl or 3.0.6-dev, etc.).

### Example (with 3.0.4-dev installed)


https://github.com/user-attachments/assets/a74f8872-4047-4307-b1c0-0d522581066d



### Related issue(s)/PR(s)
Resolves #16466